Changes to installation

+++ .config/.bin/install.sh
@@ -21,8 +21,11 @@ echo "Getting Oh-my-zsh..."
 git clone https://github.com/ohmyzsh/ohmyzsh.git ~/.oh-my-zsh
 
 echo "Getting Zsh plugins..."
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/>
-git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/>
+git clone https://github.com/zsh-users/zsh-syntax-highlighting.git \ 
+	${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+git clone https://github.com/zsh-users/zsh-autosuggestions \ 
+	${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+
 
 echo "Making zsh the default shell..."
 chsh -s $(which zsh)
@@ -30,47 +33,75 @@ chsh -s $(which zsh)
 # Remove unnecessary shit
 echo "Cleaning unnecessary programs and stuff from HOME..."
 sudo pacman -R palemoon-bin moc
-rm -rf $HOME/Public \
-	$HOME/Music \
-	$HOME/Videos \
-	$HOME/Desktop \
-	$HOME/Pictures \
-	$HOME/Templates \
-	$HOME/'.moonchild productions/' \
-	$HOME/.moc
-#rm .zshrc.pre-oh-my-zsh .bash_history .histfile
+
+rm -f .zshrc.pre-oh-my-zsh .bash_history .histfile
+
+dir="$HOME/Public $HOME/Music $HOME/Videos $HOME/Desktop $HOME/Pictures
+     $HOME/Templates $HOME/'.moonchild productions' $HOME/.moc"
+
+for d in $dir
+do
+  if [ -d $d ]; then
+      rm -rf $d
+      printf "$d deleted...\n"
+  fi
+done
 
 # Move config files to $XDG_CONFIG_HOME
 echo "Moving files to XDG_CONFIG_HOME..."
 
+XDG_CONFIG_HOME="$HOME/.config"
+XDG_CACHE_HOME="$HOME/.cache"
+XDG_DATA_HOME="$HOME/.local/share"
+
 mkdir -p ~/.config/i3
-mv ~/.i3/config ~/.config/i3
-rm -rf .i3
+
+if [ -f "~/.i3/config ]; then
+    mv ~/.i3/config ~/.config/i3
+    rm -rf .i3
+    echo "Moved i3 config to $XDG_CONFIG_HOME/i3"
+fi 
 
 mkdir -p ~/.config/gtk-2.0
-mv ~/.gtkrc-2.0 ~/.config/gtk-2.0/gtkrc
+if [ -f  "~/.gtkrc-2.0 ]; then
+    mv ~/.gtkrc-2.0 ~/.config/gtk-2.0/gtkrc
+    echo "Moved gtkrc-2.0 to $XDG_CONFIG_HOME/gtk-2.0"
+fi
 
 mkdir -p ~/.config/X11
+if [ -f ~/.xinitrc ]; then
+    mv ~/.xinitrc ~/.config/X11/xinitrc
+    echo "Moved xinitrc to $XDG_CONFIG_HOME/X11"
+fi
+
 #mv ~/.Xauthority ~/.config/X11/Xauthority
-mv ~/.xinitrc ~/.config/X11/xinitrc
 #mv ~/.Xresources ~/.config/X11/Xresources
 
-mv ~/.oh-my-zsh ~/.config/oh-my-zsh
+if [ -d ~/.oh-my-zsh ]; then
+    mv ~/.oh-my-zsh ~/.config/oh-my-zsh
+    echo "Moved Oh-my-zsh to $XDG_CONFIG_HOME/oh-my-zsh"
+fi
 
 echo "Newsboat..."
 mkdir -p "$XDG_DATA_HOME"/newsboat "$XDG_CONFIG_HOME"/newsboat
-exec newsboat
-mv ~/.newsboat ~/.config/newsboat
+if [ -d ~/.newsboat ]; then
+    mv ~/.newsboat ~/.config/newsboat
+    echo "Moved newsboat to $XDG_CONFIG_HOME/newsboat"
+fi
+#mv ~/.newsboat ~/.config/newsboat
+
 
 # Clone dotfiles using git bare repository
-git clone --bare --recursive https://github.com/mikkonurminen/.dotfiles \
+git clone --bare https://github.com/mikkonurminen/.dotfiles \
     "$HOME/.dotfiles"
 
 function dotgit() {
     /usr/bin/env git --git-dir="$HOME/.dotfiles/" --work-tree="$HOME" "$@"
 }
 
-dotgit checkout
+dotgit config --local status.showUntrackedFiles no
+
+dotgit checkout &> /dev/null
 if [ "$?" -ne 0 ]; then
     mkdir -p "$HOME/.dotfiles-backup"
     dotgit checkout 2>&1 \
@@ -80,5 +111,4 @@ if [ "$?" -ne 0 ]; then
     dotgit checkout
 fi
 
-dotgit submodule update --recursive --remote
-dotgit config --local status.showUntrackedFiles no
+#dotgit submodule update --recursive --remote
